refactor : 친구 추가 페이지 보낸요청, 받은 요청, 검색 결과 렌더링 컴포넌트로 분리
